design: 비밀번호 재발급 UI
